requirements.txt
fastmcp
lark-oapi
python-telegram-bot
python-dotenv
fastapi
uvicorn